Sparkc 472 b 2.0

+++ spark-cassandra-connector/src/main/scala/org/apache/spark/sql/cassandra/CassandraSourceRelation.scala
@@ -31,6 +31,7 @@ private[cassandra] class CassandraSourceRelation(
     tableRef: TableRef,
     userSpecifiedSchema: Option[StructType],
     filterPushdown: Boolean,
+    confirmTruncate: Boolean,
     tableSizeInBytes: Option[Long],
     connector: CassandraConnector,
     readConf: ReadConf,
@@ -53,11 +54,21 @@ private[cassandra] class CassandraSourceRelation(
 
   override def insert(data: DataFrame, overwrite: Boolean): Unit = {
     if (overwrite) {
-      connector.withSessionDo {
-        val keyspace = quote(tableRef.keyspace)
-        val table = quote(tableRef.table)
-        session => session.execute(s"TRUNCATE $keyspace.$table")
+      if (confirmTruncate) {
+        connector.withSessionDo {
+          val keyspace = quote(tableRef.keyspace)
+          val table = quote(tableRef.table)
+          session => session.execute(s"TRUNCATE $keyspace.$table")
+        }
+      } else {
+        throw new UnsupportedOperationException(
+          """You are attempting to use overwrite mode which will truncate
+          |this table prior to inserting data. If you would merely like
+          |to change data already in the table use the "Append" mode.
+          |To actually truncate please pass in true value to the option
+          |"confirm.truncate" when saving. """.stripMargin)
       }
+
     }
 
     implicit val rwf = SqlRowWriter.Factory
@@ -277,6 +288,7 @@ object CassandraSourceRelation {
       tableRef = tableRef,
       userSpecifiedSchema = schema,
       filterPushdown = options.pushdown,
+      confirmTruncate = options.confirmTruncate,
       tableSizeInBytes = tableSizeInBytes,
       connector = cassandraConnector,
       readConf = readConf,
